+++ src/Console/Command.php
@@ -19,9 +19,11 @@ namespace DatabaseBackup\Console;
 use Cake\Console\Arguments;
 use Cake\Console\BaseCommand;
 use Cake\Console\ConsoleIo;
+use Cake\Console\ConsoleOptionParser;
 use Cake\Core\Configure;
 use Cake\Datasource\ConnectionInterface;
 use Cake\Datasource\ConnectionManager;
+use Cake\Datasource\Exception\MissingDatasourceConfigException;
 use Override;
 use Symfony\Component\Filesystem\Path;
 
@@ -31,13 +33,34 @@ use Symfony\Component\Filesystem\Path;
 abstract class Command extends BaseCommand
 {
     /**
-     * Gets the `Connection`.
-     *
-     * @return \Cake\Datasource\ConnectionInterface
+     * @var \Cake\Datasource\ConnectionInterface
      */
-    public function getConnection(): ConnectionInterface
+    protected ConnectionInterface $Connection;
+
+    /**
+     * @inheritDoc
+     */
+    #[Override]
+    protected function buildOptionParser(ConsoleOptionParser $parser): ConsoleOptionParser
     {
-        return ConnectionManager::get(Configure::readOrFail('DatabaseBackup.connection'));
+        return $parser
+            ->addOptions([
+                'connection' => [
+                    'default' => 'default',
+                    'help' => __d(
+                        'database_backup',
+                        'Name of the alternative connection to use, for example if you are not using the default connection'
+                    ),
+                ],
+                'timeout' => [
+                    'help' => __d(
+                        'database_backup',
+                        'Timeout for shell commands. Default value: {0} seconds',
+                        Configure::readOrFail('DatabaseBackup.processTimeout')
+                    ),
+                    'short' => 't',
+                ],
+            ]);
     }
 
     /**
@@ -58,8 +81,16 @@ abstract class Command extends BaseCommand
     #[Override]
     public function execute(Arguments $args, ConsoleIo $io): void
     {
-        $io->out(__d('database_backup', 'Connection: {0}', $this->getConnection()->config()['name']));
-        $io->out(__d('database_backup', 'Driver: {0}', $this->getConnection()->config()['driver']));
+        try {
+            /** @var string $connectionName */
+            $connectionName = $args->getOption('connection');
+            $this->Connection = ConnectionManager::get($connectionName);
+        } catch (MissingDatasourceConfigException $E) {
+            $io->abort($E->getMessage());
+        }
+
+        $io->out(__d('database_backup', 'Connection: {0}', $this->Connection->config()['name']));
+        $io->out(__d('database_backup', 'Driver: {0}', $this->Connection->config()['driver']));
 
         if ($args->getOption('timeout')) {
             $io->verbose(