+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # 2.x branch
 ## 2.14 branch
 ### 2.14.2
+* the constructor method `AbstractBackupUtility` now accepts `$Connection` as an optional argument, as string (e.g.
+  `default` or `test`) or a `ConnectionInterface` instance. This means that by instantiating `BackupExport` or
+  `BackupImport` (which extend `AbstractBackupUtility`) it is possible to set a connection beyond the default one;
+* `ExportCommand` and `ImportCommand` now accept the `--connection` option to set a connection other than the default;
+* due to the two new introductions just described, the `DatabaseBackup.connection` configuration name is deprecated
+  and will be removed in a later release;
 * configuration names `DatabaseBackup.mysql`, `DatabaseBackup.postgres` and `DatabaseBackup.sqlite` are now deprecated
   and will be removed in a future release (backwards compatibility will be maintained up to branch `2.15.x`). Use the
   `DatabaseBackup.Mysql`, `DatabaseBackup.Postgres` and `DatabaseBackup.Sqlite` names instead in your bootstrap (the 
-  driver name begins with a capital letter).
+  driver name begins with a capital letter);
+* removed from `ExportCommand::buildOptionParser()` the reference (as a helper) to the old `send` option, which was no
+  longer present (it had been forgotten).
 
 ### 2.14.1
 * many improvements for the `AbstractExecutor` abstract class;